+++ src/Component/Provider/OpenIdConnectClient.php
@@ -4,27 +4,24 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Component\Provider;
 
+use Heptacom\AdminOpenAuth\Component\OpenIdConnect\OpenIdConnectService;
 use Heptacom\AdminOpenAuth\Service\TokenPairFactoryContract;
 use Heptacom\OpenAuth\Behaviour\RedirectBehaviour;
 use Heptacom\OpenAuth\Client\Contract\ClientContract;
 use Heptacom\OpenAuth\Struct\TokenPairStruct;
 use Heptacom\OpenAuth\Struct\UserStruct;
 use Psr\Http\Message\RequestInterface;
-use TheNetworg\OAuth2\Client\Provider\Azure;
 
-/**
- * @deprecated tag:v5.0.0 will be replaced by microsoft_azure_oidc provider
- */
-class MicrosoftAzureClient extends ClientContract
+class OpenIdConnectClient extends ClientContract
 {
     private TokenPairFactoryContract $tokenPairFactory;
 
-    private Azure $azureClient;
+    private OpenIdConnectService $openIdConnectService;
 
-    public function __construct(TokenPairFactoryContract $tokenPairFactory, array $options)
+    public function __construct(TokenPairFactoryContract $tokenPairFactory, OpenIdConnectService $openIdConnectService)
     {
         $this->tokenPairFactory = $tokenPairFactory;
-        $this->azureClient = new Azure($options);
+        $this->openIdConnectService = $openIdConnectService;
     }
 
     public function getLoginUrl(?string $state, RedirectBehaviour $behaviour): string
@@ -45,7 +42,7 @@ class MicrosoftAzureClient extends ClientContract
 
     public function refreshToken(string $refreshToken): TokenPairStruct
     {
-        return $this->tokenPairFactory->fromLeagueToken($this->getInnerClient()->getAccessToken('refresh_token', [
+        return $this->tokenPairFactory->fromOpenIdConnectToken($this->getInnerClient()->getAccessToken('refresh_token', [
             'refresh_token' => $refreshToken,
         ]));
     }
@@ -59,41 +56,29 @@ class MicrosoftAzureClient extends ClientContract
         }
 
         $token = $this->getInnerClient()->getAccessToken('authorization_code', $options);
-        $user = $this->getInnerClient()->get('me', $token);
+        $user = $this->getInnerClient()->getUserInfo($token);
 
-        $emails = [];
-
-        if (($email = \trim($user['mail'] ?? '')) !== '') {
-            $emails[] = $email;
-        }
-
-        // TODO break fallback behaviour in v5 and make it configurable
-        if (($email = \trim($user['userPrincipalName'] ?? '')) !== '') {
-            $emails[] = $email;
+        $name = sprintf('%s %s', $user->getName() ?? '', $user->getFamilyName() ?? '');
+        if (empty(trim($name))) {
+            $name = $user->getNickname() ?? $user->getEmail();
         }
 
         return (new UserStruct())
-            ->setPrimaryKey($user['objectId'])
-            ->setTokenPair($this->tokenPairFactory->fromLeagueToken($token))
-            ->setDisplayName($user['displayName'])
-            ->setPrimaryEmail(\array_pop($emails))
-            ->setEmails($emails)
-            ->setPassthrough(['resourceOwner' => $user]);
+            ->setPrimaryKey($user->getSub())
+            ->setTokenPair($this->tokenPairFactory->fromOpenIdConnectToken($token))
+            ->setDisplayName($name)
+            ->setPrimaryEmail($user->getEmail())
+            ->setEmails([$user->getEmail()])
+            ->addPassthrough('picture', $user->getPicture());
     }
 
     public function authorizeRequest(RequestInterface $request, TokenPairStruct $token): RequestInterface
     {
-        $result = $request;
-
-        foreach ($this->getInnerClient()->getHeaders($token->getAccessToken()) as $headerKey => $headerValue) {
-            $result = $result->withAddedHeader($headerKey, $headerValue);
-        }
-
-        return $result;
+        return $request->withAddedHeader('Authorization', 'Bearer ' . $token->getAccessToken());
     }
 
-    public function getInnerClient(): Azure
+    public function getInnerClient(): OpenIdConnectService
     {
-        return $this->azureClient;
+        return $this->openIdConnectService;
     }
 }