+++ src/Service/Provider/MicrosoftAzureOidcProvider.php
@@ -4,24 +4,28 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Service\Provider;
 
-use Heptacom\AdminOpenAuth\Component\Provider\MicrosoftAzureClient;
+use Heptacom\AdminOpenAuth\Component\OpenIdConnect\OpenIdConnectConfiguration;
+use Heptacom\AdminOpenAuth\Component\OpenIdConnect\OpenIdConnectService;
+use Heptacom\AdminOpenAuth\Component\Provider\OpenIdConnectClient;
 use Heptacom\AdminOpenAuth\Service\TokenPairFactoryContract;
 use Heptacom\OpenAuth\Client\Contract\ClientContract;
 use Heptacom\OpenAuth\ClientProvider\Contract\ClientProviderContract;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * @deprecated tag:v5.0.0 will be replaced by microsoft_azure_oidc provider
- */
-class MicrosoftAzureProvider extends ClientProviderContract
+class MicrosoftAzureOidcProvider extends ClientProviderContract
 {
-    public const PROVIDER_NAME = 'microsoft_azure';
+    public const PROVIDER_NAME = 'microsoft_azure_oidc';
 
     private TokenPairFactoryContract $tokenPairFactory;
 
-    public function __construct(TokenPairFactoryContract $tokenPairFactory)
-    {
+    private OpenIdConnectService $openIdConnectService;
+
+    public function __construct(
+        TokenPairFactoryContract $tokenPairFactory,
+        OpenIdConnectService $openIdConnectService
+    ) {
         $this->tokenPairFactory = $tokenPairFactory;
+        $this->openIdConnectService = $openIdConnectService;
     }
 
     public function provides(): string
@@ -33,18 +37,21 @@ class MicrosoftAzureProvider extends ClientProviderContract
     {
         return parent::getConfigurationTemplate()
             ->setDefined([
+                'tenantId',
                 'clientId',
                 'clientSecret',
                 'scopes',
                 // TODO remove in v5
                 'redirectUri',
             ])->setRequired([
+                'tenantId',
                 'clientId',
                 'clientSecret',
             ])->setDefaults([
                 'scopes' => [],
                 'redirectUri' => null,
             ])
+            ->setAllowedTypes('tenantId', 'string')
             ->setAllowedTypes('clientId', 'string')
             ->setAllowedTypes('clientSecret', 'string')
             ->setAllowedTypes('scopes', 'array')
@@ -55,6 +62,7 @@ class MicrosoftAzureProvider extends ClientProviderContract
     {
         $result = parent::getInitialConfiguration();
 
+        $result['tenantId'] = '';
         $result['clientId'] = '';
         $result['clientSecret'] = '';
 
@@ -63,6 +71,13 @@ class MicrosoftAzureProvider extends ClientProviderContract
 
     public function provideClient(array $resolvedConfig): ClientContract
     {
-        return new MicrosoftAzureClient($this->tokenPairFactory, $resolvedConfig);
+        $config = new OpenIdConnectConfiguration();
+        $config->assign($resolvedConfig);
+        $config->setDiscoveryDocumentUrl('https://login.microsoftonline.com/' . $resolvedConfig['tenantId'] . '/v2.0/.well-known/openid-configuration');
+
+        $service = $this->openIdConnectService->createWithConfig($config);
+        $service->discoverWellKnown();
+
+        return new OpenIdConnectClient($this->tokenPairFactory, $service);
     }
 }