+++ src/Service/Provider/OpenIdConnectClientEntityRepository.php
@@ -8,10 +8,7 @@ use Heptacom\AdminOpenAuth\Component\Provider\DefaultInjectingEntityRepositoryDe
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 
-/**
- * @deprecated tag:v5.0.0 will be replaced by microsoft_azure_oidc provider
- */
-class MicrosoftAzureClientEntityRepository extends DefaultInjectingEntityRepositoryDecorator
+class OpenIdConnectClientEntityRepository extends DefaultInjectingEntityRepositoryDecorator
 {
     protected function injectDefaults(array $payload, Context $context): array
     {
@@ -22,7 +19,7 @@ class MicrosoftAzureClientEntityRepository extends DefaultInjectingEntityReposit
         }
 
         foreach ($payload as &$item) {
-            if (($item['provider'] ?? '') !== MicrosoftAzureProvider::PROVIDER_NAME) {
+            if (($item['provider'] ?? '') !== OpenIdConnectProvider::PROVIDER_NAME) {
                 continue;
             }